Fixing Database schema for contacts table and in process of removing gorm

+++ server/models/model.go
@@ -3,7 +3,7 @@ package models
 import "time"
 
 type Model struct {
-	ID        uint      `json:"id,omitempty" yaml:"id,omitempty" gorm:"primarykey"`
+	ID        uint64    `json:"id,omitempty" yaml:"id,omitempty" gorm:"primarykey"`
 	CreatedAt time.Time `json:"created_at,omitempty" yaml:"created_at,omitempty"`
 	UpdatedAt time.Time `json:"-" yaml:"-"`
 }
